+++ src/components/Notifications.vue
@@ -2,7 +2,7 @@
 <div :class="{'toast-notifications-list-pf': toast}" v-show="$slots.default || notifications.length">
   <slot/>
   <pf-notification ref="notification" v-for="(n, i) in notifications" :key="i" :persistent="n.persistent" :type="n.type" :toast="toast" @dismiss="dismiss(i)">
-    <div v-html="n.message"></div>
+    <span v-html="n.message"></span>
   </pf-notification>
 </div>
 </template>
@@ -18,10 +18,8 @@ export default {
   },
 
   props: {
-    toast: {
-      type: Boolean,
-      default: false,
-    },
+    toast: Boolean,
+    persistent: Boolean,
   },
 
   data() {
@@ -31,12 +29,12 @@ export default {
   },
 
   methods: {
-    add(notification, type = 'info', persistent = false) {
+    add(notification, type = 'info', persistent) {
       if (typeof notification != 'object') {
         notification = {
           message: notification,
           type: type,
-          persistent: persistent,
+          persistent: typeof persistent == 'undefined' ? this.persistent : persistent,
         };
       }
       this.notifications.push(notification);